python-12312sd-main/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==2.0.23
alembic==1.7.3
python-dotenv==0.18.0
pydantic==1.9.0
psycopg2
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6